site/requirements.txt
Flask==1.1.2
python-dotenv==0.15.0
cassandra-driver==3.24.0